Move files in test
