+++ tests/tests/bare_game.c
