+++ tests/utils/silent_log.c
