+++ tests/utils/silent_log.h
